LPD-20634 Move the PortletPreferences TableUpdater to the kernel base class since we need to use in different places
